--- a/src/Skia/Avalonia.Skia/SKFontCache.cs
+++ b/src/Skia/Avalonia.Skia/SKFontCache.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Concurrent;
+using SkiaSharp;
+
+namespace Avalonia.Skia
+{
+    /// <summary>
+    /// Cache for SKFonts.
+    /// </summary>
+    internal class SKFontCache : SKCacheBase<SKFont, SKFontCache>
+    {
+        
+    }
+}