--- a/src/Skia/Avalonia.Skia/SKTextBlobBuilderCache.cs
+++ b/src/Skia/Avalonia.Skia/SKTextBlobBuilderCache.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Concurrent;
+using SkiaSharp;
+
+namespace Avalonia.Skia
+{
+    /// <summary>
+    /// Cache for SKTextBlobBuilder.
+    /// </summary>
+    internal class SKTextBlobBuilderCache : SKCacheBase<SKTextBlobBuilder, SKTextBlobBuilderCache>
+    {
+
+    }
+}