## What does the pull request do?
Continuing my investigation into performance of the composer loop.  Currently, there are several places which `SKTextBlobBuilder`, `SKRoundRect` and `SKFont` objects are instanced and either fall out of scope or are disposed..  These objects are all backed by SKIA counterparts which must be disposed of.  This PR seeks to reduce how many of these items are instanced and discarded and instead reuse the instances for later.

Found out that SKTextBlobBuilder is designed to be re-used.  Once the [make](https://api.skia.org/classSkTextBlobBuilder.html#abf5e20208fd5656981191a3778ee5fef) method is called, it is reset for reuse.  Apparently SkaiSharp uses the [make method internally](https://github.com/mono/SkiaSharp/blob/a863ca3ed6b420a86f7e81afec8eba79dfda495d/binding/Binding/SKTextBlob.cs#L285).

I took the opportunity to also update the `SKPaintCache` introduced in #9441 to use the same new base caching system.

## What is the current behavior?
`SKTextBlobBuilder`, `SKRoundRect` and `SKFont`s are created and either explicitly disposed or GCed.  This leads to additional objects which are created and needlessly disposed.

## What is the updated/expected behavior with this PR?
There are new `SKTextBlobBuilderCache`, `SKRoundRectCache` and `SKFontCache` types which allow for caching and resetting of these base types for reuse.

This PR focuses on improving following methods:
#### `Avalonia.Skia.PlatformRenderInterface.CreateGlyphRun(IGlyphTypeface, Double, IReadOnlyList, Point)`
Now uses ` SKFontCache` for caching.
Now uses ` SKTextBlobBuilderCache` for caching,

#### `Avalonia.Skia.DrawingContextImpl.DrawRectangle(IBrush, IPen, RoundedRect, BoxShadows)`
Now uses ` SKRoundRectCache` for caching,

#### `Avalonia.Skia.TextShaperImpl.ShapeText(ReadOnlyMemory, TextShaperOptions)`
Now has an internal static cache on the `TextShaderImpl` class which caches currently used languages.
ShapeText

## How was the solution implemented (if it's not obvious)?
Added caching classes and utilized where possible.

## Benchmarks
| Type     | Method                                                                                             | N         | Time (ms) | Memory (MB) |
|----------|----------------------------------------------------------------------------------------------------|-----------|-----------|-------------|
| Baseline | Avalonia.Skia.PlatformRenderInterface.CreateGlyphRun(IGlyphTypeface, Double, IReadOnlyList, Point) |  250,000  | 426       | 91          |
| PR       | Avalonia.Skia.PlatformRenderInterface.CreateGlyphRun(IGlyphTypeface, Double, IReadOnlyList, Point) |  250,000  | 92        | 45          |
|          |                                                                                                    |           |           |             |
| Baseline | Avalonia.Skia.DrawingContextImpl.DrawRectangle(IBrush, IPen, RoundedRect, BoxShadows)              |  250,000  | 550       | 27          |
| PR       | Avalonia.Skia.DrawingContextImpl.DrawRectangle(IBrush, IPen, RoundedRect, BoxShadows)              |  250,000  | 90        | 0           |
|          |                                                                                                    |           |           |             |
| Baseline | Avalonia.Skia.TextShaperImpl.ShapeText(ReadOnlyMemory, TextShaperOptions)                          |  250,000  | 744       | 169         |
| PR       | Avalonia.Skia.TextShaperImpl.ShapeText(ReadOnlyMemory, TextShaperOptions)                          |  250,000  | 246       | 177**       |

All benchmarks are run with this [Benchmark](https://github.com/DJGosnell/Avalonia/tree/render-benchmark) which renders 500 Square Rectangles, 500 Round Rectangles and 500 simple texts every frame until 500 frames have passed.  When the test takes less time, it means that the rendering took less time to complete. the same 500 frames.

** Currently Reviewing why this usage is higher...

Baseline Benchmark Time **10,235ms**
![BaselineOverall](https://user-images.githubusercontent.com/224169/218226028-b67d3787-a135-4012-af4f-06afe1dbf628.png)

PR Benchmark Time **7,419ms**
![PROverall](https://user-images.githubusercontent.com/224169/218226029-d099275d-840f-4311-ab10-413c6664c1ca.png)


<details><summary>DotTrace Details</summary>
<p>

CreateGlyphRun_Baseline
![CreateGlyphRun_Baseline](https://user-images.githubusercontent.com/224169/218222484-132632e1-03d1-4690-b097-3100cccfffb6.png)

CreateGlyphRun_PR
![CreateGlyphRun_PR](https://user-images.githubusercontent.com/224169/218222486-30b5d735-ebe2-4be2-ad76-1b6eefc36f7d.png)

DrawRectangle_Baseline
![DrawRectangle_Baseline](https://user-images.githubusercontent.com/224169/218222488-3797fe01-3333-4194-b020-b37409dcdb50.png)

DrawRectangle_PR
![DrawRectangle_PR](https://user-images.githubusercontent.com/224169/218222489-484cc3f3-9794-4ddc-aa46-7303a2256030.png)

Language_Baseline
![Language_Baseline](https://user-images.githubusercontent.com/224169/218223606-5a949a97-5cc4-4484-9c5e-e2d3e45d0162.png)

Language_PR
![Language_PR](https://user-images.githubusercontent.com/224169/218223610-8cb70168-eacd-4e42-a69e-07405625aef8.png)

ShapeText_Baseline
![ShapeText_Baseline](https://user-images.githubusercontent.com/224169/218225638-1d576c69-7166-454e-b1a9-8b57e2d2efb4.png)

ShapeText_PR
![ShapeText_PR](https://user-images.githubusercontent.com/224169/218225639-39a2fc83-95d5-4499-b97b-884761eaa693.png)

</p>
</details>

## Checklist

- [x] Added unit tests (if possible)? [Benchmark](https://github.com/DJGosnell/Avalonia/tree/render-benchmark)
- [x] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Documentation with user documentation

## Breaking changes
None.

## Obsoletions / Deprecations
N/A

## Fixed issues
N/A

## Notes
It may be worth a discuss about these locations the caches are used and if there is any possibility for cross thread usage.  If not, we can replace the ConcurrentBag<T> with a Queue<T> for additional savings.