@maxkatz6 2023-02-11
> if there is any possibility for cross thread usage. If not, we can replace the ConcurrentBag with a Queue for additional savings.

Currently we support only a single rendering thread. But I think there were discussion to support more than one in the future.