@kekekeks 2023-02-11
Graphics backend can be used from multiple threads already.
1) RenderTargetBitmap is available from the UI thread in immediate mode
2) It's possible to manually create a render target from something like a framebuffer