@Gillibald 2023-02-11
Are you sure that SKTextBlobBuilder can be shared by multiple threads? We should definitely verify that FormattedText can be created on any thread.