@DJGosnell 2023-02-13
> Are you sure that SKTextBlobBuilder can be shared by multiple threads? We should definitely verify that FormattedText can be created on any thread.

@Gillibald Just want to make sure that I understand what is going and how my caching strategy works in this case. 
-  All the caches require you to manually add back the cached item.  This has two effects: 
    1. There is no overhead for any wrapper classes and handling disposal of wrappers.
    2. If for some reason the item does not get added back to the cache, it will get properly disposed like any normal CLR object. 
- Any time a cached item is not in the cache, a new one is created even if they are on the same thread.
- No time between requesting a cached item and returning it to the cache, should the cached object be used in multiple places across the codebase.  Meaning the same cached object will not be used on another thread nor another place on the same thread.  If one thread requests an object and another does simultaneously, they will both get separate objects whether they be from the cache or new instances created by the cache.

Now that I explained my cache implementation a bit better, is the following flow the concern?

  1. Cached item for SKTextBlobBuilder is requested, used and returned on one thread.
  2. Cached item for the same SKTextBlobBuilder is requested and used on another thread.

Through this implementation of the cache, no single SKTextBlobBuilder will be used simultaneously.