@Gillibald 2023-02-14
As long as no SKTextBlobBuilder-related resources are bound to a dedicated thread this is fine.