@DJGosnell 2023-02-15
Per the [documentation](https://api.skia.org/classSkTextBlobBuilder.html#abf5e20208fd5656981191a3778ee5fef), the `make` method _Resets [SkTextBlobBuilder](https://api.skia.org/classSkTextBlobBuilder.html) to its initial empty state, allowing it to be reused to build a new set of runs._

I've run extensive testing to ensure that SKTextBlobBuilder is can be used used across threads (not simultaneously) and one of the methods I used was to create a multi-threaded benchmark which has the baseline current method of using the SKTextBlobBuilder (instancing per usage) and PR caching usage.  See the benchmarks and traces below.

I never ran into any exceptions or unexpected behavior throughout testing.

https://github.com/DJGosnell/SKTextBlobBuilderBenchmarks

|   Method |     Mean |    Error |   StdDev | Ratio | RatioSD |     Gen0 |    Gen1 | Allocated | Alloc Ratio |
|--------- |---------:|---------:|---------:|------:|--------:|---------:|--------:|----------:|------------:|
| Baseline | 33.14 ms | 1.293 ms | 3.813 ms |  1.00 |    0.00 | 187.5000 |       - |   1.65 MB |        1.00 |
|       PR | 17.45 ms | 0.455 ms | 1.343 ms |  0.53 |    0.07 | 140.6250 | 15.6250 |   1.05 MB |        0.64 |

**Baseline**
![Baseline](https://user-images.githubusercontent.com/224169/219118013-de831278-491a-4a8a-8afd-e7707dcbc93a.png)

**PR**
![PR](https://user-images.githubusercontent.com/224169/219118046-b63e7caf-8a7e-4217-be81-85246c0ce10f.png)