@avaloniaui-team 2023-02-15
You can test this PR using the following package version. `11.0.999-cibuild0030075-beta`. (feed url: https://pkgs.dev.azure.com/AvaloniaUI/AvaloniaUI/_packaging/avalonia-all/nuget/v3/index.json) [PRBUILDID]